03_multi_RAG_agents_handsoff_no_tool/requirements.txt
# Microsoft Agent Framework (preview)
agent-framework --pre
agent-framework-azure-ai --pre

# Orchestrations (handoff, sequential, concurrent, etc.)
agent-framework-orchestrations --pre

# Azure SDKs
azure-identity

# Local vector store
chromadb

# Web scraping & text processing
requests
beautifulsoup4
lxml
